delegate state seperation 1
